move upload-file implementation to browser/ and node/ folders
