+++ test/node/upload-file.test.js
@@ -1,9 +1,9 @@
-/* eslint-env mocha */
+/* eslint-env jest */
 'use strict'
 
-const { expect } = require('chai')
+const fs = require('fs')
 const nock = require('nock')
-const uploadFile = require('../../src/utils/upload-file-node')
+const uploadFile = require('../../src/node/upload-file')
 
 const UPLOAD_URL = 'https://upload-url.example.com'
 
@@ -20,15 +20,19 @@ describe('#uploadFile', () => {
 
   it('returns empty body', () => {
     const fileMimeType = 'text/html'
-    const file = new window.File([''], 'filename', { type: fileMimeType })
+    const stats = fs.statSync('jest.config.js')
+    const fileSizeInBytes = stats.size
+    const readStream = fs.createReadStream('jest.config.js')
+
     const promise = uploadFile({
       uploadUrl: `${UPLOAD_URL}/uploads`,
       fileMimeType: fileMimeType,
-      file: file
+      contentLength: fileSizeInBytes,
+      file: readStream
     })
 
     promise.then(body => {
-      expect(body).to.eql({})
+      expect(body).toEqual({})
     })
   })
 })
